refactor(sandbox): move native sandbox logic into dedicated backend module
Move all macOS sandbox-exec and Linux bwrap wrapping logic from sandbox.ts
into backends/native.ts implementing the SandboxBackend interface. sandbox.ts
becomes a thin orchestration layer that delegates to NativeBackend.

Part of #2021.
